rrr: convert rpc/inmemory_channel to module (rrr.inmemory_channel)
inmemory_channel.cpp module + .hpp shim. The in-memory channel
backend (deterministic-tests transport, no kernel/socket I/O). Uses
rrr.channel for the ChannelConnection/Listener/Factory bases,
rrr.threading for SpinMutex, rrr.logging for Log_error. No reactor
dependency — that's why it could land independently of the reactor
cluster.

37 modules. 80.87 s wallclock.
